only 'helm dep build' if renovatebotwrapper was changed

+++ .github/helm-dep.sh
@@ -1,20 +1,24 @@
 #!/bin/bash
 set -ex
 
-cd ./lab/renovatebotwrapper
+if git diff --name-only | grep lab/renovatebotwrapper ; then
 
-helm repo add argo https://argoproj.github.io/argo-helm
-helm repo add coredns https://coredns.github.io/helm
-helm repo add istio https://istio-release.storage.googleapis.com/charts
-helm repo add jetstack https://charts.jetstack.io
-helm repo add metrics-server https://kubernetes-sigs.github.io/metrics-server/
-helm repo add mojo2600 https://mojo2600.github.io/pihole-kubernetes/
-helm repo add openfaas https://openfaas.github.io/faas-netes/
-helm repo add traefik https://traefik.github.io/charts
+  cd ./lab/renovatebotwrapper
 
-# helm dep update
-helm dep build
+  helm repo add argo https://argoproj.github.io/argo-helm
+  helm repo add coredns https://coredns.github.io/helm
+  helm repo add istio https://istio-release.storage.googleapis.com/charts
+  helm repo add jetstack https://charts.jetstack.io
+  helm repo add metrics-server https://kubernetes-sigs.github.io/metrics-server/
+  helm repo add mojo2600 https://mojo2600.github.io/pihole-kubernetes/
+  helm repo add openfaas https://openfaas.github.io/faas-netes/
+  helm repo add traefik https://traefik.github.io/charts
 
-cd ./charts
-ls -1 *.tgz | xargs -t -I{} -P1 tar xvf {}
-rm -v *.tgz
+  # helm dep update
+  helm dep build
+
+  cd ./charts
+  ls -1 *.tgz | xargs -t -I{} -P1 tar xvf {}
+  rm -v *.tgz
+
+fi
